fix(review): expose exploration records at package gateway
